Document and Move Filter Code
1. Move all filter code to common/dsp/filters
2. Document the filter architecture in QuadFilterChain.h

Closes #3241
